Move the vtables from `value_type` to a new registry that is allocated for each value_trait
this resolves various safety issues and removes some indirections.  Now to deref we just need to find the static vtabletregistry for the trait (static dispatch) lookup the vtable based on the value type (map lookup), then all we need to do is compose the fat pointer.

Compared to the previous approach we are saving an allocation per trait impl and 2 memory references (due to the ValueType lookup and the Box derefence), we are also saving the various comparisions required for the downcasts.