[dynamicIO] Use filled Resume Data Cache for final-phase prerenders (#79743)
When a route with dynamic params is prerendered using defined params
from `generateStaticParams`, the filled Resume Data Cache is now used
for the final-phase prerender of the matching optional fallback shell.

This ensures that the fallback shell can reuse existing cache entries,
which partially mitigates the performance hit we accepted by splitting
up the prerendering into two phases.

Furthermore, it will allow us to short-circuit fallback params access in
`"use cache"` functions. Instead of having to wait for the timeout
error, we will be able to regard cache misses (excluding cached pages
and layouts, whose params access is already handled swiftly) to be
caused by including the params in the cache key, if
`allowEmptyFallbackShell` is `true`.