[dynamicIO] do not continue prerender after error in prospective render
When the prospective prerender errors we curretly ignore the result and continue through. Since modules can error this can lead to situations where the final render assumes all modules are already loaded when they are not and this can lead to misattribution of dynamicIO reasons. We don't normally retry renders and we don't need to do so in this case either.