[dynamicIO] Don't abort prospective render on sync access during SSR
When filling caches we never want to let sync IO interrupt the cache filling render. We already handled this for the react-server render where most caches live. But modules are a form of caches and they similarly rely on the prospective render to "fill" their entry. Prior to this change during SSR prerenders a sync IO could cause the prospective render to abort early. This led to the case where the final render would also potentially abort early for something like a serial sync IO in module scope.