[dynamicIO] Track module loading in dev always
In dev we prerender and render concurrently. The former to validation dynamicIO rules. The latter to stream a response as soon as possible. If we only track module loading during prerenders it is possible to kick off loading a chunk during hte render and then not observe that chunk load during prerender leading to an early abort before all modules have had a chance to load.

To solve this we simply always track module loading in dev.