Add next-intlayer to i18n solution list on app router
Next-intlayer is an i18n solution that adopts a per-component approach to managing translations. 

It’s been added to the i18n page-router documentation, but is fully compatible with the App Router as well.

This commit adds next-intlayer to the list of i18n solutions for the Next.js App Router.