Enable `@typescript-eslint/switch-exhaustiveness-check` rule
Enabling the `@typescript-eslint/switch-exhaustiveness-check` eslint
rule, ensures that we don't accidentally forget covering existing cases,
and it allows us to introduce new types with confidence that we'll
handle them everywhere they are used. This is useful in general, but
especially important when we introduce new work unit store types, which
we're about to do.

This rule requires type information, so we're adding it to
`.eslintrc.cli.json`, which also means that problems reported by this
rule won't be shown in the IDE by default.