Simplify running test apps locally with `ppr` or `dynamicIO` enabled
We want to respect the `__NEXT_EXPERIMENTAL_PPR` and
`__NEXT_EXPERIMENTAL_CACHE_COMPONENTS` environment variables in the test
apps, so that we can run them locally with these features enabled
without needing to modify the config files.

So we're removing the requirement that `__NEXT_TEST_MODE` needs to be
set at the same time. Usually you don't want to set this variable
because it also changes other behavior in a few places.

This really has no downside, because it was already possible for users
to set both variables and affect the flags, and now this can be done
with a single variable. It allows us to keep the noise out of the test
app's next config files.

We're also moving the assignments from the default config into
`enforceExperimentalFeatures` to get the nice output when printing the
experimental features in the CLI.