[Segment Cache] Support dynamic head prefetching
Fixes an issue where opting into dynamic prefetching with prefetch={true}
would not apply to head data (like the title), only the page data.
Although the head was being sent by the server as part of the prefetch
response, it wasn't being transferred correctly to the prefetch cache.

The net result is that you can now fully prefetch a page with a dynamic
title without any additional network requests on navigation.