Properly handle hanging promise rejections during prerendering
We are using hanging promises for excluding request-specific data from a
prerender. When prerendering is aborted, we reject all hanging promises.
In dev mode, React uses these rejections to construct more detailed
owner stacks. This allows us to provide better errors during the
prerender validation.

For this to work reliably, we need to ensure that all hanging promises
are aborted after React's rendering is aborted. However, some hanging
promises are created before React starts rendering (when creating the
RSC component tree). Furthermore, for memory reasons, the first hanging
promise adds a single abort listener that's shared with all subsequent
hanging promises (see #77480). In this case, the shared abort listener
for rejections is added before React can add its own abort listener.

When using the same abort signal for both the React render and the
hanging promises, the hanging promises would be rejected too early,
preventing React from properly constructing the owner stacks.

To ensure that the order of operations is correct, we now use a separate
abort signal for the hanging promises that is only aborted after the
render signal is aborted. For parity, we're doing this not only in dev
mode but also in production mode, so that the hanging promise rejection
handling is always the same.

Reverts #81501.