[turbopack] Reuse shared Promise for loaded chunks
This is still problematic when React sees this special
Promise for the first time since it'll still need
a microtask to find out it's already resolved.

The proper fix is to fully instrument the Promise
in Turbopack (`status`) field so that React can make
use of the `status` field instead of instrumenting
it separately.