[router-act] Fix redirects not being intercepted
Fixes a problem with the internal testing helper `createRouterAct` where
a redirected response was not being intercepted by the Playwright
listener, and would fail if you tried to assert on its contents. I
hadn't noticed this before because most tests don't assert on the
contents of the initial tree prefetch, only the subsequent data
requests; it only became a problem once I started inlining data into the
tree prefetch.

Hendrik's fix in https://github.com/vercel/next.js/pull/80967
addressed a related issue where a redirect caused a race condition, for
the same reason: Playwright treats requests that happen as the result
of a redirect differently from "normal" requests and does not pass them
to the `route.page` handler. `on('request')` doesn't seem to suffer
from the same problem, so we use that instead.

The reason we can't use `on('request')` for _all_ requests is that it's
read-only — you can't intercept the request and control its fulfillment,
which is necessary for the "block" option. I have yet to figure out a
way to make that work, so for now, "block" won't work for redirected
responses (I made it error instead, so at least it doesn't silently
fail). Not a big deal in the meantime, though, since this is just an
internal testing helper and that particular testing pattern is easy to
work around.