[router-act] Response may include multiple matches
This removes an assertion check that fails if the same response includes
multiple expected substrings. I can't recall my original thinking for
adding that error, but I don't think it's desirable; as much as
possible, our tests should not rely on the number of requests being
made under the hood. I noticed this while working on segment inlining,
since that causes some segments to be bundled into the same response.