[Segment Cache] Implement prefetch inlining
In the Segment Cache implementation, we split page prefetches into
multiple requests so that each one can be cached and deduped
independently. For example, if multiple pages share the same layout, the
data for that layout only needs to be fetched once. This approach can
greatly reduce the overall transfer size at the cost of additional
requests, which is typically minimal, especially since it happens during
the prefetch phase.

However, some segments are small enough that the potential caching
benefits are not worth additional network overhead. So we can inline
them directly into the initial route tree prefetch response, potentially
avoiding a separate request. Again, it's not a perfect trade-off, since
that separate request may have been skipped anyway if the segment was
already cached. It's just a heuristic. But below some size, inlining
is worth it.

(Publicly, we may end up referring to this strategy as prefetch
_outlining_ rather than _inlining_, since compared to the current
Next.js behavior, that's the part that's actually new in the
Segment Cache.)

I've chosen 2KB (gzipped) as the default threshold. We can do some more
research and tweak this later.

In addition to using the size of the segment prefetch, we also take into
consideration the total inlined size of the tree prefetch. The idea is
that even if each individual segment is below the inlining threshold,
at some point it no longer makes sense to add more bytes to the tree
prefetch, because the tree prefetch requests are unique per URL and
cannot be deduped the way segments can. I've chosen 10KB as the default
for this limit.

Deeper segments are more likely to be unique to a page, so they are
more likely to benefit from inlining. We prioritize the deepest segments
when inlining.

For now, the prefetch thresholds are hardcoded. We may make them
configurable, by ideally the defaults are good enough that it isn't
necessary. One reason to expose the configuration, though, is that some
apps may be more sensitive to the number of requests than others. By
setting a very large inlining threshold, you can effectively disable
the layout deduping feature and "revert" back to something closer to
the existing Next.js behavior. We'll think about this more
before release.

Another thing worth considering is whether to use different inlining
thresholds for different network conditions. For example, mobile devices
on slow networks may benefit from more aggressive inlining, whereas
devices on WiFi benefit more from deduping. To do this, we would
generate multiple prefetch responses per page and the client would
request the appropriate one based on the
[Network Information API](https://developer.mozilla.org/en-US/docs/Web/API/Network_Information_API).