Change the `rcstr!` macro to avoid lazy locks
This removes branches from the allocation path and instead we just need to tag the pointer on each call.  To do this we needed PrehashedString to be const compatible which required some indirection to allow it to just directly store static strings.  This will add a bit of overhead to `as_str`, `eq` and `drop` but speed up allocation.