[CacheComponents] Use fallback params when validating dynamic routes in dev
When validating a dynamic route in dev with Cache Components we must not use more params than would be know at build timebecause we want to ensure dev acts as an appropriate debug environment for build validation issues. Any time a validation fails for reasons other than fallback params it would have failed as well with fallback params. So we find the smallest set of fallback params that would be used during the build (which might be empty if you provide a complete set through generateStaticParams) and we use that.