Avoid duplicate WebSocket connection for global error pages
When we're rendering a global error page (something with `<html
id="__next_error__">`), we are wrapping the root element in a component
that renders a top-level dev overlay to show the global error. The
function that's wrapping this element was previously creating an
unnecessary second WebSocket connection to handle the case where the
error is removed by editing a server component, and the page is
subsequently reloaded. However, this can be handled by the first
WebSocket connection that's created in the root element just fine, we
just need to ensure that the initial value for
`RuntimeErrorHandler.hadRuntimeError` is set based on the document
element's ID, which is the same condition we use for rendering the
wrapper.

This change is covered by existing tests, e.g.
`test/development/app-dir/missing-required-html-tags/index.test.ts`.