[Segment Cache] Optimize successive lookups
This is the start of a larger refactor of the data type used to store
Segment Cache entries. In this first step, I've focused on optimizing
successive cache lookups, by storing a backpointer from the cache entry
to the internal map node that references it. Subsequent operations can
then use this backpointer directly to skip the map lookup.

This requires imposing a constraint that regular JS maps do not have:
a value cannot be stored at multiple keypaths simultaneously. This is
fine, though, because the larger goal of this refactor is to introduce
a more robust mechanism for reusing cache entries across different keys,
similar to the "fallback PPR" mechanism we use on the server. This will
be introduced in subsequent PRs.