[Segment Cache] Generalize fallback implementation
Adds `getWithFallback` to the internal cache map type. This is a
generalized mechanism for storing "fallback" cache entries. A fallback
entry is one that does not vary by a particular key, and can therefore
be reused for any concrete value of that key. We intend to use this
strategy much more broady in the future, but we already do this in a
few places.

One existing place where we already have fallback behavior is
interception routes. If the server instructs us that a route will
never be intercepted, we reuse the route entry for all possible
referer URLs (confusingly named the "Next-Url" internally).

This PR implements `getWithFallback` and updates the cache to use it
for interception route fallback behavior.