fix: aliased navigations should apply scroll handling (#82900)
When routing to the same page with different searchParams w/ a loading
segment, we have handling to skip RSC roundtrip and re-use the existing
RSC data. However, this handling was missing logic that exists in
`navigate-reducer` which sets the `scrollableSegments`. As a result, the
router would stay at the existing scroll position rather than scrolling
to the top of the page, as it would in the case of a non-aliased
navigation.

Fixes NEXT-4679