Unhook WebSocket
Currently, the WebSocket that's used for `next dev` is created in a
React hook inside of the `HotReload` component, which is (transitively)
rendered by the `ServerRoot` component, but only after the initial RSC
response is processed, at least partially.

This creates a problem when integrating React's debug channel, because
the root model of the RSC payload will be blocked until the associated
debug info has been received through the debug channel. Since we will be
using the existing WebSocket connection for the debug channel, we need
to establish it before the `ServerRoot` component is rendered.

The next-devtools dispatcher already queues events, before React has
provided the `dispatch` function in `DevOverlayRoot`. So creating the
WebSocket outside of rendering is already supported in that regard.