feat(route-types): allow override for route handlers and layouts
When we enforce the type of a dynamic segment and we want to use this type in components, it's currently doable only for App pages.
To enable it for route handlers and layouts, we need to add `& any` to the props sent to the components/handlers.

Example with a route `/[locale]/abc` where we only allow the `en` & `fr` locales.

```tsx
type Locale = "en" | "fr"

type PageComponent<AppRoute extends AppRoutes> = ComponentType<
  PageProps<AppRoute> extends infer Props
    ? Props extends { params: Promise<infer PageParams> }
      ? PageParams extends ParamMap[AppRoute]
        ? PageParams extends Record<"locale", unknown>
          ? PageParams["locale"] extends string
            ? Props & { params: Promise<PageParams & { locale: Locale }> }
            : never
          : never
        : never
      : never
    : never
>

const MyPage: PageComponent<"/[locale]/abc"> = async ({ params }) => {
  const { locale } = await params // locale has the following type: "en" | "fr"
  ...
}
```