Update caching.mdx
In the docs it says => ( The default caching behavior of fetch (e.g., when the cache option is not specified) is equal to setting the cache option to no-store )

But when I code, I found that fetch default behaviour and setting the cache option to no-store are different.

After building process ....

When changing routes. default is stored as cache. I see that because there is no new network request when I change route. In setting the cache option to no-store, it is making a new network request as soon as i get back to the page.

so docs and real working one are mismatch.

Please build... and check
Change routes...