Use the automatic runtime for type checking
TypeScript understands 2 types of JSX runtimes. The classic runtime
looks at the `React.JSX` namespace in the file. The automatic runtime
looks at the `JSX` namespace exported from `react/jsx-runtime` or
`react/jsx-dev-runtime` respectively.

The following `jsx` options use the automatic runtime:
- `react-jsx`
- `react-jsx-dev`

The following `jsx` options use the classic runtime:
- `preserve`
- `react`
- `react-native`

The difference between `preserve` and `react`, is that `preserve` does
not consider the JSX factory to be used as a value, only for types. This
means that although `react` will throw a type error for access to the
`React` UMD global, `preserve` will not.

Note that this is about types, not values. Although React 19 removed the
UMD global, my suggestion to remove the UMD global from the React 19
types was rejected.

Since Next.js compiles JSX to use the automatic runtime, the correct
`jsx` option is `react-jsx`.

The practical reason I ran into this, is because the `preact/compat`
types also define the `React` UMD global. This causes conflicts if
`preact/compat` is used with Next.js. In my case we pulled it in as a
transient dependency.