[test] Improve debug logs for Playwright tests
This commit enhances the debug logging for Playwright tests by prefixing
the messages with a timestamp, which helps debugging timing related
issues. It also sends debug messages directly to Node's stdout,
bypassing Jest's verbose output.

In addition, we're not running the tests in silent mode anymore in CI.
This is a relic from the past, and we do already set
`TRACE_PLAYWRIGHT=true`, which indicates that we want to see all the
debug output.