Don't execute inventory_submit in rust analyzer
This is triggering an bug "expected integer or floating pointer number after '-'".

We don't really understand the issue (perhaps related to a rustc nightly release?) but as far as we can tell it is just triggered by the transitive expansion of `inventory::submit!`.  So we can just skip doing that with rust_analyzer since none of that code affects types.