Move lazy eviction logic to cache-map module
Anything to do with retrival, lifetimes, or eviction needs to go in the
cache-map.ts module because it affects the fallback algorithm. For
example, when performing a lookup, if an entry is stale, it needs to be
treated as semantically equivalent to if the entry was not present at
all. Similarly, cache-map.ts needs to be aware of the "revalidating"
field because if an entry is stale, but there's a revalidating entry,
the revalidating entry should be swapped in for stale entry. We would
get a different result if we were to wait to upsert the entry until
after the lookup was complete.

If there's logic that's not related to the fallback algorithm, though,
we should prefer to put it in cache.ts.