Turbopack: don't define `process.cwd()` in node_modules (#83452)
Don't compile-time-define `process.cwd()` inside of node_modules.

This means that node_modules cannot break outside of their individual packages anymore, so in the worst case, a given package includes all of itself, but not anymore the whole project.

This means that a `readFileSync(path.join(process.cwd(), "something.txt")))` won't cause that file to be listed in the NFT file anymore.