[CI] Fix GitHub Actions matrix configuration for React versions
The `build-and-test` workflow runs on `canary` are currently annotated
with the following error:

```
Error when evaluating 'strategy' for job 'test-rspack-production'. .github/workflows/build_and_test.yml (Line: 412, Col: 13): Matrix exclude key 'react' does not match any key within the matrix
```

This addresses the issue for `test-rspack-production`, and also fixes
inconsistencies in other jobs regarding the handling of this matrix
dimension.