[Cache Components] Allow sync IO inside console methods
console methods may be patched to do things like add a timestamp to the log row. While we can't determine the meaning of readin the current time in arbitrary contexts we can know with reasonable certainty that any sync IO inside a console method is not being used in the logical output of the program. To account for this we can make a special affordance for sync IO inside of these methods that allows them to run without interrupting the prerender early.

Originally I wanted to make it so that even if you patch these methods in userland we can still exit the context but it turns out this is very complex and difficult because we have many patches layered together and they need to run in the correct order. So essentially this just permits sync IO in the preexisting implementations of the console methods. This is probably sufficient since typically these methods aren't patched in userspace but in the runtime host itself. If we recieve feedback that this is insufficient we can consider alternative later.

The implementation moves the console dimming logic to it's own AsyncLocalStorage so we can exit the workUnitAsyncStorage without altering the console dimming behavior. Conceptually you can think of the internal implementation of console methods as existing outside of any particular render (prerender or otherwise).