Conditionally collect affecting sources
These are only needed for tracing, not normal resolution so we can save some time and memory tracking these values.  By doing this we also increase cachability as we are no longer tracking this unneeded data.

I considered modeling it as an `Option<Vec<ResolvedVc<Box<dyn Source>>>>`, but empty vecs are just as good memory wise and this approach is simpler.