Allow `metadataBase` to be a string URL in addition to `URL` instance
As of #49343, in dev mode only, we accidentally change the
`metadataBase` from an `URL` instance to a URL string, when passing it
to child layouts or pages via the `parent` argument in
`generateMetadata`. This is because we clone the metadata object via
`JSON.stringify`/`parse` and then freeze it, to prevent relying on
mutating resolved metadata. The replacer/reviver functions we used to
clone the metadata didn't properly handle preserving `URL` instances as
they intended, so they were turned into strings.

Because user code already had to handle the case of `metadataBase`
potentially being a string at runtime in dev mode, we can just
officially support that as well in our types, not only for the
`ResolvedMetadata` input type, but also for the `Metadata` output type.

This also enables `'use cache'` being used in `generateMetadata`,
because React currently does not support transporting `URL` instances
from one environment to another (e.g. `Server` to `Cache`).

Note that all other URL properties in `Metadata` and `ResolvedMetadata`
are already `string | URL`, so this only affects `metadataBase`.