[Segment Cache] Re-implement refresh reducer
This re-implements the refresh reducer to use the Segment Cache-style
flow, similar to how the navigate reducer already works.

Most of the logic can be reused from what was already implemented for
navigations. The bulk of the changes are related to the case where
a "default" parallel route slot is not present on the current page, and
so needs to be refreshed from an older page's URL.

Theoretically, this should be the last remaining case where we rely on
the lazy fetching logic in LayoutRouter. I think there may still be some
error cases where we still rely on that, though. Once these are all
addressed, we can remove the lazy-fetching mechanism entirely.