Remove unneeded fetch delay
We added a fetch delay during prerendering so that if a fetch was cached (even if it had not explicit cache configuration) it would not be available in the final prerender pass's static phase. This should not be necessary anymore since we now halt the fetch even in the warmup render.