Simplify Metadata components
Metadata and Viewport litter the component tree with a number of Next.js internal implementation details. We can't yet get rid of all of these but we can simplify by consolidating these components or eliminating them altogether.

One change is to render the metadata as a promise that resolves direclty to the client tags that must be in the head. Another is to render a promise that will directly reject with the error reason rather than reconstructing a digest manually.

In the process I removed a bunch of module indirection too