[test] Unflake static image test
[flakiness metric](https://app.datadoghq.com/ci/test/runs?query=test_level%3Atest%20%40git.repository.id%3A%22github.com%2Fvercel%2Fnext.js%22%20%40test.name%3A%22Static%20Image%20Component%20Tests%20development%20mode%20Should%20allow%20an%20image%20with%20a%20static%20src%20to%20omit%20height%20and%20width%22%20%40test.type%3A%22turbopack%22%20%40test.status%3A%22fail%22&agg_m=count&agg_m_source=base&agg_t=count&currentTab=overview&eventStack=&fromUser=false&index=citest&start=1759057503838&end=1759662303838&paused=false)

I suspect that the test `Should allow an image with a static src to omit
height and width` is frequently failing with `waiting for
locator('#basic-static') to be visible` because the image is covered by
one of the images with a `fill` prop, which are rendered later in the
page. This means that the image is not actually visible, even though it
is in the DOM. Playwright then fails to find it. However, I wasn't able
to reproduce this locally.

The fix is to wrap the `fill` images in a relatively positioned div with
a set height, so that they don't cover the other images.